[UPDATE] Add new assets
