+++ lib/support/utils/app_assets.dart
@@ -1,0 +1,23 @@
+class AppAssets {
+  AppAssets._();
+
+  static const String _assetsPath = 'assets/images/';
+
+  //Images
+  static const String loginCover = '${_assetsPath}login_cover.png';
+  static const String championshipCover = '${_assetsPath}home_championship_card.png';
+  static const String nbaCover = '${_assetsPath}home_nba_card.png';
+  static const String esportsCover = '${_assetsPath}home_redbull_card.png';
+  static const String xBetLogo = '${_assetsPath}x_bet_logo.png';
+  static const String betSafeLogo = '${_assetsPath}bet_safe_logo.png';
+  static const String betssonLogo = '${_assetsPath}betsson_logo.png';
+
+  //Icons
+  static const String icGoogle = '${_assetsPath}ic_google.png';
+  static const String icApple = '${_assetsPath}ic_apple.png';
+  static const String icImperioLogo = '${_assetsPath}imperio_logo.png';
+  static const String icChampionsLeague = '${_assetsPath}ic_champions.png';
+  static const String icEuroLeague = '${_assetsPath}ic_euro_league.png';
+  static const String icSerieA = '${_assetsPath}ic_serie_a.png';
+  static const String icLaLiga = '${_assetsPath}ic_la_liga.png';
+}
