--- lib/support/utils/style/app_assets.dart
@@ -1,12 +1,0 @@
-class AppAssets {
-  AppAssets._();
-
-  static const String _assetsPath = 'assets/images/';
-
-  //Images
-  static const String loginCover = '${_assetsPath}login_cover.png';
-
-  //Icons
-  static const String icGoogle = '${_assetsPath}ic_google.png';
-  static const String icApple = '${_assetsPath}ic_apple.png';
-}
