--- lib/support/utils/style/app_colors.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class AppColors {
-  AppColors._();
-
-  static const Color white = Color.fromRGBO(252, 252, 252, 1);
-  static const Color lightGrey = Color.fromRGBO(230, 230, 230, 1);
-  static const Color black = Color.fromRGBO(0, 0, 0, 1);
-  static const Color darkGrey = Color.fromRGBO(85, 82, 82, 1);
-  static const Color ultraDarkGrey = Color.fromRGBO(40, 44, 42, 1);
-  static const Color yellow = Color.fromRGBO(245, 215, 10, 1);
-}
