--- lib/support/utils/style/app_theme.dart
@@ -1,27 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'app_colors.dart';
-
-class AppTheme {
-  AppTheme._();
-
-  static final theme = ThemeData(
-    scaffoldBackgroundColor: AppColors.white,
-
-    /// ElevatedButtonThemeData
-    elevatedButtonTheme: ElevatedButtonThemeData(
-      style: ElevatedButton.styleFrom(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(24)),
-        minimumSize: const Size.fromHeight(56),
-      ),
-    ),
-
-    /// OutlinedButtonThemeData
-    outlinedButtonTheme: OutlinedButtonThemeData(
-      style: OutlinedButton.styleFrom(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(24)),
-        minimumSize: const Size.fromHeight(56),
-      ),
-    ),
-  );
-}
